Moved `experiment` directory to `service`
